abstract out openfast reader from generic container
